in Controllers: annotation CrossOrigin: in AdminController changed companies without coupons to no coupons.  and made a new function for without coupons for getting the companies without getting coupons from DB.
In Company added a constructor for query used in findAllCompaniesWithoutCoupons() in CompanyRepository.
In Coupon added a constructor for adding image_url to random coupons.

In CustomerController and CouponRepository added a query to select coupons which are not of the specific customer add with amount > 0.
I used "Not in" in the query because of MySql, in place of exclude/minus in normal SQL.